requirements.txt
ldap3>=2.9.1
mysql-connector-python>=8.3.0
pyasn1>=0.4.8
